Extract ActionContext and dispatcher from main.js; add E.COMMAND_ISSUED
- js/action-context.js: new module with buildActionContext() and
  initActionDispatcher(ctx) — both are now testable without DOM or
  game state initialization
- js/events.js: add E.COMMAND_ISSUED ("command:issued") event
- js/graph.js: export fitGraph(cy) — removed from main.js
- js/log-renderer.js: subscribe to COMMAND_ISSUED to log "> cmd" entries
- js/console.js: emit COMMAND_ISSUED in submitCommand() instead of
  calling addLogEntry() directly; subscribe to COMMAND_ISSUED to call
  pushHistory() — removes duplicate history.unshift() in keydown handler
- js/types.js: remove logCommand from ActionContext typedef
- js/main.js: use buildActionContext() + initActionDispatcher(ctx) + import
  fitGraph from graph.js — shrinks from 131 to 75 lines
- tests/node-actions.test.js: drop logCommand from mockCtx
- Makefile: add action-context.js to lint target

143/143 tests passing.
